+++ R/CFAxisLongitude.R
@@ -46,7 +46,7 @@ CFAxisLongitude <- R6::R6Class("CFAxisLongitude",
       }
 
       if (is.null(rng)) {
-        if (length(self$values) > 1L) {
+        if (length(private$values) > 1L) {
           ax <- self$clone()
           ax$group <- group
           ax
@@ -54,10 +54,10 @@ CFAxisLongitude <- R6::R6Class("CFAxisLongitude",
           .make_scalar(1L)
       } else {
         if (rng[1L] == rng[2L])
-          .make_scalar(self$values[rng[1L]])
+          .make_scalar(private$values[rng[1L]])
         else {
           dim <- NCDimension$new(-1L, self$name, rng[2L] - rng[1L] + 1L, FALSE)
-          lon <- CFAxisLongitude$new(group, var, dim, self$values[rng[1L]:rng[2L]])
+          lon <- CFAxisLongitude$new(group, var, dim, private$values[rng[1L]:rng[2L]])
           bnds <- self$bounds
           if (inherits(bnds, "CFBounds")) lon$bounds <- bnds$sub_bounds(group, rng)
           private$copy_label_subset_to(lon, idx)